Moved some API functions to WrittenBook, fixed/implemented inserting of pages.
